requirements.txt
# Core libraries
numpy
pandas

# Financial data analysis
pyodbc
sqlalchemy
cx_Oracle
psycopg2-binary
arctic
web3

# Plotting and visualization
matplotlib
plotly
bokeh
dash
dash-bootstrap-components

# Geospatial analysis
geopandas

# Machine learning
scikit-learn

# Sentiment analysis
textblob

# API and web services
flask
requests

# C++ integration
pybind11

# Testing
pytest
pytest-cov

# Additional dependencies
inflect
